Comprehensive backend update: sistema completo de despachos, cajas variables, documentos tributarios y mejoras WMS
✨ Nuevas funcionalidades:
- Sistema completo de despachos con picking y delivery tracking
- Gestión avanzada de cajas variables con lotes FIFO
- Documentos tributarios (boletas/facturas) con control SII
- Sistema WMS de enrolamiento para carne
- Alertas de productos sin precio
- Precios de proveedores para cajas variables
- Tipos de venta (unitario/peso) integrados

🔧 Mejoras técnicas:
- Cantidad decimal en items_pedido (float)
- Locales con tipos específicos (venta/bodega)
- Control de estados SII para documentos
- Gemini Vision API para OCR avanzado
- Sistema de puntos optimizado
- Múltiples migraciones aplicadas

📦 Archivos agregados:
- Routers: despachos, stock_cajas, configuracion, alertas
- Schemas: despacho, enrolamiento, gemini_vision
- Scripts: seeding y validación de deploy
- Documentación: deploy y multi-instancia
- Tests: integración y validación

🛠️ Fixes y optimizaciones:
- Relationships en stock_cajas corregidas
- Endpoints de maestras optimizados
- Seguridad mejorada en autenticación
- Validaciones de tipos de pedido

+++ schemas/item_pedido.py
@@ -9,20 +9,20 @@ class ItemPedidoBase(BaseModel):
     """Schema base de ItemPedido."""
     pedido_id: int = Field(..., gt=0)
     producto_id: int = Field(..., gt=0)
-    cantidad: int = Field(..., gt=0)
+    cantidad: float = Field(..., gt=0)  # Cambiado a float para soportar decimales
     precio_unitario_venta: float = Field(..., gt=0)
 
 
 class ItemPedidoCreate(BaseModel):
     """Schema para crear un ItemPedido (sin pedido_id, se asigna automáticamente)."""
     producto_id: int = Field(..., gt=0)
-    cantidad: int = Field(..., gt=0)
+    cantidad: float = Field(..., gt=0)  # Cambiado a float para soportar decimales
     precio_unitario_venta: float = Field(..., gt=0)
 
 
 class ItemPedidoUpdate(BaseModel):
     """Schema para actualizar un ItemPedido."""
-    cantidad: Optional[int] = Field(None, gt=0)
+    cantidad: Optional[float] = Field(None, gt=0)  # Cambiado a float para soportar decimales
     precio_unitario_venta: Optional[float] = Field(None, gt=0)
